Allow using JDK 1.7 to compile faban.
In 1.7, specifying only the target version as was done in faban
build.xml is an error. If we specify source version instead, it
implies target and compilation succeeds. Doing it this way works with
JDK 1.6 as well (tested these changes with 1.6.0_21 and 1.7.0_25).

(Not sure if we really need to be compiling this to 1.5 anymore, but
I've left that unchanged for now. Can be upgraded later if desired.)

Also, build.xml was passing the same version to rmic 'stubversion'
which is incorrect (though it only generates a warning). The only
valid values for stubversion are 1.1 and 1.2, it doesn't take the full
range of JDK versions. Changed this to "1.2" which is the default.