Make registry activity log more info.
Mismatches in how a service is registered vs. looked up is often a
cause of trouble/confusion when setting up faban. This diff adds info
logging when services are both registered and looked up. This can make
it easier to diagnose related problems (because info logs show up in
the UI log view of a run).