fix issue where service jars were dropped from RMI classPath
CmdAgentImpl uses a cache of service jars. The previous logic would only
add the service jars to allClassPath if they were not in the cache but
the same method replaces the content of allClassPath at the end so it
effectly removed the service jars from the classPath after the second
invocation. This change uses the same cache but includes the service
jars after each invocation.